fix bug when reading a broken package.json + refactor
